feat: page 컨벤션 변경  #33, #34, #35, #36
